[refactoring] refactoring fcm register feature
