+++ app/src/main/java/com/nsd/talk/ui/register/RegisterActivity.kt
@@ -1,20 +1,28 @@
 package com.nsd.talk.ui.register
 
 import android.Manifest
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
+import android.os.Build
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import android.view.animation.AlphaAnimation
 import android.view.animation.AnimationSet
 import android.view.animation.DecelerateInterpolator
 import android.widget.Toast
 import androidx.activity.viewModels
+import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.messaging.FirebaseMessaging
 import com.nsd.talk.databinding.ActivityRegisterBinding
 import com.nsd.talk.ui.main.MainActivity
 
-
+@RequiresApi(Build.VERSION_CODES.O)
 class RegisterActivity : AppCompatActivity() {
     private val binding by lazy { ActivityRegisterBinding.inflate(layoutInflater) }
     private val REQUEST_PERMISSIONS = 1
@@ -27,7 +35,7 @@ class RegisterActivity : AppCompatActivity() {
     }
 
     private fun isFirstStartApp() {
-        if(!viewModel.isFirstStartApp(applicationContext)) {
+        if (viewModel.isFirstStartApp(applicationContext)) {
             viewModel.setFirstStartAppPreference(applicationContext)
             val intent = Intent(this@RegisterActivity, MainActivity::class.java)
             startActivity(intent)
@@ -45,9 +53,11 @@ class RegisterActivity : AppCompatActivity() {
                 llPhoneNumber.visibility = View.VISIBLE
                 btnRegister.visibility = View.VISIBLE
                 btnNext.visibility = View.INVISIBLE
+                viewModel.name = etName.text.toString()
             }
         }
         btnRegister.setOnClickListener {
+            viewModel.phoneNumber = etPhoneNumber.text.toString()
             checkPermission()
         }
     }
@@ -87,10 +97,37 @@ class RegisterActivity : AppCompatActivity() {
                     ).show()
                     finish()
                 } else {
-                    val intent = Intent(this@RegisterActivity, MainActivity::class.java)
-                    startActivity(intent)
+                    register()
                 }
             }
         }
+    }
+
+    private fun register() {
+        firebaseSetUp()
+        val intent = Intent(this@RegisterActivity, MainActivity::class.java)
+        startActivity(intent)
+    }
+
+    private fun firebaseSetUp() {
+        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
+            if (!task.isSuccessful) {
+                Log.w("[MainActivity]", "Fetching FCM registration token failed", task.exception)
+                return@OnCompleteListener
+            }
+            val token = task.result
+            Log.d("MainActivity", token)
+            viewModel.register(token)
+        })
+        createNotificationChannel("1111", "fcm")
+    }
+
+    private fun createNotificationChannel(id: String, name: String) {
+        val importance = NotificationManager.IMPORTANCE_DEFAULT
+        val channel = NotificationChannel(id, name, importance)
+
+        val notificationManager: NotificationManager =
+            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        notificationManager.createNotificationChannel(channel)
     }
 }
